Taking in events from GLFW- processing window close properly! Updated premake to VS2022.

+++ Zephyr/src/Zephyr/Events/MouseEvent.h
@@ -14,7 +14,7 @@ namespace Zephyr
 	class ZEPHYR_API MouseMove : public MouseEvent
 	{
 	public:
-		MouseMove(Vec2<int> pos)
+		MouseMove(Vec2<float> pos)
 			:pos(pos) {}
 
 		inline Vec2<float> GetPos() const
